Remove dependency for composer installers
The dependency for composer installers is unnecessary at this level. 
The project that includes wp-statuses should have composer installers required - but that is not the job for this plugin to ask for it. 

This requirement brings my setup in trouble as this plugins composer.json requires a very specific composer installer version that is outdated.